+++ backend/src/main/java/wooteco/retrospective/domain/dao/TimeDao.java
@@ -1,0 +1,12 @@
+package wooteco.retrospective.domain.dao;
+
+import wooteco.retrospective.domain.attendance.Time;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface TimeDao {
+    List<Time> findAll();
+    int update(Long timeId, Time time);
+    Optional<Time> findById(long id);
+}
